Add code for color conversion functions
The following color conversion functions are added:
- RGB2GRAY
- GRAY2RGB
- RGB2YUV
- YUV2RGB

Only RGB2GRAY have an example and a test associated with it right now. The other
functions are NOT yet tested, but there's no harm for checking them in now since
they're not used anywhere. I also expect that the general logic should be
correct, if there're any bugs it'd be minor things like the types used are not
correct, the coefficients are not exactly correct, there're typos, etc. We can
fix these easily later.