Add test for RGB2GRAY
This test is an extension of examples/color_conversion/rgb2gray. The extra
part is running the functions with a 16-bit image.

Right now this code runs both the 8 bit and 16 bit width. For the end goal, we
want this code to be configurable e.g. it can run either 8-bit or 16-bit based
on a parameter. We also don't really want to use dg.exp because this code is
exposed to the users.
The idea so far is to have a bash script that runs a loop through different
pixel types e.g. 8UC1, 16UC1. But we still need to figure out more details about
how we want to do this.