Add a simple frame buffer control to the ddr_access_wrapper.
Both DDR writer and reader HLS modules take in a base address in DDR for write
and read.  Previously the base addresses are always tied to the same value,
so writer and reader are accessing the same frame simultaneously.  This can
cause the read side to read a "frame" that contains pixels from different
frames (because a new frame is being written while read is happening).

This commit adds a simple FrameBufferControl module such that the writer and
reader will access different frames.
The FrameBufferControl module provides the base addresses for the writer and
reader, and updates the addresses when the writer finishes writing a frame
(takes writer_finish as input).
Each time the DDR writer finishes writing a frame to DDR, FrameBufferControl
provides a new address for the DDR writer to write the next frame, and
provides to the DDR reader with the address of latest frame (the frame that
just got written to DDR by DDR writer).
The number of frames and each frame's byte size can be parameterized.

Also do some refactoring to `debayer.hpp` and `format_conversions.hpp`.