Added test agc+brightness+color demo
The demo now has a PASS/FAIL check against a golden image. Even though the demo is meant to be used for video, a test with a single image can at least check if the verilog code would work for one frame.

The check also has a flag (SMALL) to select a small image for quick cosim testing. If this flag is not set (Default) a 4K image is used.